Move request/response encoders/decoders to the versioned `rpc` packages
