+++ cosmos-server/src/it/scala/com/mesosphere/cosmos/PackageSearchSpec.scala
@@ -1,10 +1,9 @@
 package com.mesosphere.cosmos
 
 import cats.data.Xor.Right
-import com.mesosphere.cosmos.circe.Decoders._
-import com.mesosphere.cosmos.circe.Encoders._
 import com.mesosphere.cosmos.http.MediaTypes
-import com.mesosphere.cosmos.model._
+import com.mesosphere.cosmos.rpc.v1.circe.Decoders._
+import com.mesosphere.cosmos.rpc.v1.circe.Encoders._
 import com.mesosphere.cosmos.rpc.v1.model.{SearchRequest, SearchResponse, SearchResult}
 import com.mesosphere.cosmos.test.CosmosIntegrationTestClient.CosmosClient
 import com.mesosphere.universe.v2.model.{Images, PackageDetailsVersion, ReleaseVersion}
