+++ cosmos-server/src/main/scala/com/mesosphere/cosmos/rpc/v1/circe/MediaTypedEncoders.scala
@@ -1,0 +1,43 @@
+package com.mesosphere.cosmos.rpc.v1.circe
+
+import com.mesosphere.cosmos.circe.DispatchingMediaTypedEncoder
+import com.mesosphere.cosmos.http.MediaTypes
+import com.mesosphere.cosmos.rpc.v1.circe.Encoders._
+import com.mesosphere.cosmos.rpc.v1.model._
+
+object MediaTypedEncoders {
+
+  implicit val capabilitiesEncoder: DispatchingMediaTypedEncoder[CapabilitiesResponse] =
+    DispatchingMediaTypedEncoder(MediaTypes.CapabilitiesResponse)
+
+  implicit val packageListEncoder: DispatchingMediaTypedEncoder[ListResponse] =
+    DispatchingMediaTypedEncoder(MediaTypes.ListResponse)
+
+  implicit val packageListVersionsEncoder: DispatchingMediaTypedEncoder[ListVersionsResponse] =
+    DispatchingMediaTypedEncoder(MediaTypes.ListVersionsResponse)
+
+  implicit val packageDescribeV1Encoder: DispatchingMediaTypedEncoder[DescribeResponse] =
+    DispatchingMediaTypedEncoder(MediaTypes.V1DescribeResponse)
+
+  implicit val packageInstallEncoder: DispatchingMediaTypedEncoder[InstallResponse] =
+    DispatchingMediaTypedEncoder(MediaTypes.InstallResponse)
+
+  implicit val packageRenderEncoder: DispatchingMediaTypedEncoder[RenderResponse] =
+    DispatchingMediaTypedEncoder(MediaTypes.RenderResponse)
+
+  implicit val packageRepositoryAddEncoder: DispatchingMediaTypedEncoder[PackageRepositoryAddResponse] =
+    DispatchingMediaTypedEncoder(MediaTypes.PackageRepositoryAddResponse)
+
+  implicit val packageRepositoryDeleteEncoder: DispatchingMediaTypedEncoder[PackageRepositoryDeleteResponse] =
+    DispatchingMediaTypedEncoder(MediaTypes.PackageRepositoryDeleteResponse)
+
+  implicit val packageRepositoryListEncoder: DispatchingMediaTypedEncoder[PackageRepositoryListResponse] =
+    DispatchingMediaTypedEncoder(MediaTypes.PackageRepositoryListResponse)
+
+  implicit val packageSearchEncoder: DispatchingMediaTypedEncoder[SearchResponse] =
+    DispatchingMediaTypedEncoder(MediaTypes.SearchResponse)
+
+  implicit val packageUninstallEncoder: DispatchingMediaTypedEncoder[UninstallResponse] =
+    DispatchingMediaTypedEncoder(MediaTypes.UninstallResponse)
+
+}
